scraper/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
confection==0.1.4
cryptography==42.0.7
cymem==2.0.8
deep-translator==1.11.4
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
fpdf==1.7.2
gensim==4.3.2
google-trans-new==1.1.9
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.4.2
libretranslatepy==2.1.1
lxml==5.2.2
mathparse==0.1.2
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pdfdocument==4.0.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
Pint==0.23
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==3.13.0
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.7.5
python-dotenv==1.0.1
pytz==2024.1
PyYAML==3.13
regex==2024.5.15
reportlab==4.2.0
requests==2.31.0
rfc3986==1.5.0
scipy==1.13.0
selenium==4.20.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.2.19
sqlparse==0.5.0
srsly==2.4.8
tqdm==4.66.4
translate==3.6.1
translationstring==1.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
webdriver-manager==4.0.1
wrapt==1.16.0
wsproto==1.2.0
